Implement a proper hashCode method for graphs
This means we can finally use .distinct instead of .toSet.toList to
remove duplicates in spanningTrees.

This took a whole lot of investigation, and seems obvious now: we need
hashCode for distinct to work, and hashCode needs to deal with basic
tuples and lists (Edge and Node types are specific to instances of Graph
because of how Scala inner classes work), and it needs to be order-
independent by converting lists to sets because edges and nodes aren't
ordered in any particular way.

Well, a few things were learned about equality and hash codes, at least.