Finish the connectedComponents implementation
Broke this down a lot and ended up getting distracted by the equality
and hashCode problems addressed in the last change, but it's finally
working :)

Apologies for the distinctly un-functional approach in the collectR
function. It could most likely be solved by a foldl or an inner
recursive function, but that would have been slightly hideous and this
seemed a clearer expression of the intent (and also I'm feeling lazy).