Add basic class setup for some pathfinding experiments
Add a GridMaze, which is essentially a grid of booleans with 'true'
representing a wall and 'false' a space. These will be easy to set up
by reading string representations.

The Traversable trait is what the pathfinding experiments will use to
generically deal with whatever sort of graph we need. I've also added a
TraversableList for the sake of having another impl.