Prevent duplicate node inspections and verify on a more difficult map
The new map has a big wall obstructing a clear path through the centre,
and the test tries to find a path from the left side to the right.

As part of this I made the TraversableList return the nodes both before
and after the node whose neighbours are being requested, to add more
possibility for problems if the algorithm didn't handle duplicate
inspections correctly.