Trying out a dependency injection approach
Inspired by
http://jonasboner.com/real-world-scala-dependency-injection-di/.

This is the 'cake pattern', in which abstract component traits are
responsible for producing concrete pieces conforming to traits (eg.
LoggingComponent provides a Logger implementation) and concrete
implementations of these component traits can 'depend' on other
component traits by self-typing. We can choose a concrete set of wirings
by creating a concrete version of the component we want and having it
extend concrete versions of all the component traits on which it
depends.